+++ src/material-table.js
@@ -2,7 +2,6 @@ import React from 'react';
 import { debounce } from 'debounce';
 import equal from 'fast-deep-equal/react';
 import * as CommonValues from './utils/common-values';
-import cloneDeep from 'lodash/cloneDeep';
 import {
   Box,
   Table,
@@ -66,7 +65,10 @@ export default class MaterialTable extends React.Component {
       },
       () => {
         if (this.isRemoteData()) {
-          this.onQueryChange(this.state.query);
+          this.onQueryChange({
+            ...this.state.query,
+            page: this.props.options.initialPage || 0
+          });
         }
         /**
          * THIS WILL NEED TO BE REMOVED EVENTUALLY.
@@ -94,37 +96,29 @@ export default class MaterialTable extends React.Component {
           : '';
     }
 
-    const columnsCopy = cloneDeep(props.columns);
-
+    const savedColumns = {};
     if (props.options.persistentGroupingsId) {
       let materialTableGroupings = localStorage.getItem(
         'material-table-groupings'
       );
-
       if (materialTableGroupings) {
         materialTableGroupings = JSON.parse(materialTableGroupings);
 
         if (materialTableGroupings[props.options.persistentGroupingsId]) {
           materialTableGroupings[props.options.persistentGroupingsId].forEach(
             (savedGrouping) => {
-              const column = columnsCopy.find(
-                (col) => col.field === savedGrouping.field
-              );
-              if (column) {
-                if (!column.tableData) {
-                  column.tableData = {};
-                }
-                column.tableData.groupOrder = savedGrouping.groupOrder;
-                column.tableData.groupSort = savedGrouping.groupSort;
-                column.tableData.columnOrder = savedGrouping.columnOrder;
-              }
+              savedColumns[savedGrouping.field] = {
+                groupOrder: savedGrouping.groupOrder,
+                groupSort: savedGrouping.groupSort,
+                columnOrder: savedGrouping.columnOrder
+              };
             }
           );
         }
       }
     }
 
-    this.dataManager.setColumns(columnsCopy, prevColumns);
+    this.dataManager.setColumns(props.columns, prevColumns, savedColumns);
     this.dataManager.setDefaultExpanded(props.options.defaultExpanded);
     this.dataManager.changeRowEditing();
 
@@ -161,7 +155,9 @@ export default class MaterialTable extends React.Component {
         props.options.initialPage ? props.options.initialPage : 0
       );
     isInit && this.dataManager.changePageSize(props.options.pageSize);
-    this.dataManager.changePaging(props.options.paging);
+    this.dataManager.changePaging(
+      this.isRemoteData() ? false : props.options.paging
+    );
     isInit && this.dataManager.changeParentFunc(props.parentChildData);
     this.dataManager.changeDetailPanelType(props.options.detailPanelType);
   }
@@ -309,6 +305,10 @@ export default class MaterialTable extends React.Component {
           disabled: !!this.dataManager.lastEditingRow,
           onClick: () => {
             this.dataManager.changeRowEditing();
+            if (this.state.showAddRow) {
+              this.props.editable.onRowAddCancelled &&
+                this.props.editable.onRowAddCancelled();
+            }
             this.setState({
               ...this.dataManager.getRenderState(),
               showAddRow: !this.state.showAddRow
@@ -678,13 +678,18 @@ export default class MaterialTable extends React.Component {
         .then((result) => {
           query.totalCount = result.totalCount;
           query.page = result.page;
+          const nextQuery = {
+            ...query,
+            totalCount: result.totalCount,
+            page: result.page
+          };
           this.dataManager.setData(result.data);
           this.setState(
             {
               isLoading: false,
               errorState: false,
               ...this.dataManager.getRenderState(),
-              query
+              query: nextQuery
             },
             () => {
               callback && callback();
@@ -1064,7 +1069,6 @@ export default class MaterialTable extends React.Component {
 
   render() {
     const props = this.getProps();
-
     return (
       <DragDropContext
         onDragEnd={this.onDragEnd}